Learned the use of props extensively
This is a simple travel blog website. In this project the data was stored in data.js file . It was imported and passed as props to the react components. Understood maps method for components. Understood destructuring of the props object and different ways of passing props. Used the concept of flexbox extensively in this project using Tailwind CSS